feat: 로그인/회원가입 코드 리팩토링 및 소셜 로그인 기본 코드 제거
