+++ src/app/(route)/sign/_components/Sign.tsx
@@ -1,26 +1,10 @@
 "use client";
 
-import { Suspense, useEffect, useState } from "react";
+import { Suspense } from "react";
 import Login from "./Login";
 import Signup from "./Signup";
-import { useForm } from "react-hook-form";
-import { useRouter, useSearchParams } from "next/navigation";
-import useSign from "@/utils/UseSign";
-import { useQueryClient } from "@tanstack/react-query";
-import useAuthCheck from "@/_hooks/useAuthCheck";
-import { useSignupStore } from "@/utils/Store";
-import { useSocialStore } from "@/utils/Store";
-import useEditUserData from "@/_hooks/useEditUserData";
-import useReissue from "@/_hooks/useReissue";
-import { useApiMutation } from "@/_hooks/query";
-
-interface FormData {
-  username: string;
-  password: string;
-  email: string;
-  tel: string;
-  nickname: string;
-}
+import { useSearchParams } from "next/navigation";
+import { useRouter } from "next/navigation";
 
 interface Tabs {
   id: "login" | "signup";
@@ -29,129 +13,14 @@ interface Tabs {
 
 function Sign() {
   const router = useRouter();
-  const queryClient = useQueryClient();
-  const { data: authCheckData, isSuccess: authCheckIsSuccess } = useAuthCheck();
-  const { mutate: reissue } = useReissue();
-  const { signStateStore, setClearSignupStore } = useSignupStore();
-  const { social, resetSocial } = useSocialStore();
-  const [loginSignupState, setLoginSignupState] = useState<"login" | "signup">(
-    "login"
-  );
-  const { mutate: fetchEditUserData } = useEditUserData();
-  const [successAgree, setSuccessAgree] = useState(false);
   const searchParams = useSearchParams();
-  const statusParam = searchParams.get("status");
-  const emailParam = searchParams.get("email");
-  const {
-    mutate: fetchSign,
-    isPending,
-    isError,
-  } = useSign(loginSignupState === "login" ? "login" : "api/me/create");
-
-  const { mutate: changeStatus } = useApiMutation(
-    "put",
-    "api/members/status",
-    {},
-    {
-      headers: {
-        // Authorization: localStorage.getItem("accessToken"),
-      },
-      withCredentials: true,
-    }
-  );
-
-  const {
-    register,
-    handleSubmit,
-    getValues,
-    setValue,
-    watch,
-    reset,
-    formState: { errors },
-  } = useForm<FormData>({
-    defaultValues: {
-      email: emailParam || "",
-    },
-  });
-
-  useEffect(() => {
-    reset();
-  }, [loginSignupState]);
-
-  useEffect(() => {
-    const tokenExists = localStorage.getItem("accessToken");
-
-    if (statusParam === "PENDING" && !tokenExists) {
-      reissue(undefined, {
-        onSuccess: (data) => {
-          localStorage.setItem("accessToken", data.headers.authorization);
-        },
-      });
-    }
-  }, [statusParam]);
-
-  useEffect(() => {
-    if (social !== "" && statusParam === "PENDING") {
-      setLoginSignupState("signup");
-    }
-  }, [social]);
+  const loginSignupState = searchParams.get("sign") || "login";
 
   const tabs: Tabs[] = [
     { id: "login", label: "로그인" },
     { id: "signup", label: "회원가입" },
   ];
 
-  const onSubmit = (formData: FormData) => {
-    if (loginSignupState === "signup" && !successAgree) {
-      return alert("필수약관에 동의해주세요.");
-    }
-    if (social !== "" && loginSignupState === "signup") {
-      fetchEditUserData(
-        {
-          email: formData.email,
-          tel: formData.tel,
-          nickname: formData.nickname,
-        },
-        {
-          onSuccess: () => {
-            changeStatus({
-              email: emailParam,
-              status: "ACTIVE",
-            });
-            setSuccessAgree(false);
-            setClearSignupStore();
-            resetSocial();
-            reset();
-            router.push("/");
-          },
-          onError: (error) => {
-            console.log("error", error);
-          },
-        }
-      );
-      return;
-    }
-
-    const formDataRequest =
-      loginSignupState === "login"
-        ? { username: formData.username, password: formData.password }
-        : formData;
-    fetchSign(formDataRequest, {
-      onSuccess: (data) => {
-        if (loginSignupState === "login") {
-          localStorage.setItem("accessToken", data.headers.authorization);
-          queryClient.invalidateQueries({ queryKey: ["authCheck"] });
-          router.push("/");
-        } else {
-          setLoginSignupState("login");
-          setSuccessAgree(false);
-        }
-      },
-    });
-  };
-
-  const shouldRenderSignup =
-    signStateStore !== "" || loginSignupState !== "login";
   const loginSignupStyle =
     "w-1/2 flex items-center justify-center rounded-t-[5px] cursor-pointer border-gray-600 border-[#303030] text-[#424242]";
   return (
@@ -165,33 +34,13 @@ function Sign() {
                 ? "border-[1px] border-b-0 font-[700]"
                 : "border-b text-[#A6A6A6] border-[#A6A6A6]"
             }`}
-            onClick={() => setLoginSignupState(id)}
+            onClick={() => router.push(`/sign/?sign=${id}`)}
           >
             {label}
           </div>
         ))}
       </div>
-      <form className="w-full mt-[24px]" onSubmit={handleSubmit(onSubmit)}>
-        {shouldRenderSignup ? (
-          <Signup
-            register={register}
-            getValues={getValues}
-            isPending={isPending}
-            setSuccessAgree={setSuccessAgree}
-            formErrors={errors}
-            isError={isError}
-          />
-        ) : (
-          <Login
-            register={register}
-            setValue={setValue}
-            watch={watch}
-            isPending={isPending}
-            isError={isError}
-            formErrors={errors}
-          />
-        )}
-      </form>
+      {loginSignupState === "login" ? <Login /> : <Signup />}
     </div>
   );
 }
