+++ src/app/(route)/sign/types/login.ts
@@ -1,0 +1,7 @@
+export interface LoginFormData {
+  username: string;
+  password: string;
+  email: string;
+  tel: string;
+  nickname: string;
+}
