+++ src/app/(route)/sign/types/signup.ts
@@ -1,0 +1,6 @@
+export interface SignupFormData {
+  password: string;
+  email: string;
+  tel: string;
+  nickname: string;
+}
