+++ src/app/_components/Input.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { UseFormRegister } from "react-hook-form";
+import { RegisterOptions, UseFormRegister } from "react-hook-form";
 
 interface InputProps {
   label?: string;
@@ -13,7 +13,7 @@ interface InputProps {
   register: UseFormRegister<any>;
   isDisabled?: boolean;
   defaultValue?: string;
-  validation?: string;
+  validation?: RegisterOptions;
   required?: boolean;
 }
 
@@ -31,11 +31,8 @@ const Input: React.FC<InputProps> = ({
 }) => {
   const inputStyle = `h-[${height}px] border-[1px] py-[16px] px-[12px] rounded-[5px] text-[#181818] leading-[22px] font-[500] text-[14px] placeholder-[#A6A6A6]`;
   const labelStyle = "text-[14px] leading-[22px] text-[#424242]";
-  const helpTextStyle = "text-[14px] text-[#A6A6A6] leading-[22px] px-[16px]";
   const isDisabledInputStyle =
     inputStyle + " bg-[#EEEEEE] border-[#DBDBDB] text-[#A6A6A6]";
-  const isDisabledHelpTextStyle = helpTextStyle + " text-[#A6A6A6]";
-  const isErrorTextStyle = helpTextStyle + " text-[#D1504B]";
   return (
     <div className={`flex flex-col gap-[${gap}px]`}>
       <label htmlFor={id} className={labelStyle}>
@@ -45,7 +42,7 @@ const Input: React.FC<InputProps> = ({
       <input
         type={type}
         id={id}
-        {...register(id, { required: validation })}
+        {...register(id, validation)}
         placeholder={placeholder}
         className={isDisabled ? isDisabledInputStyle : inputStyle}
         disabled={isDisabled}
